requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
google-cloud-bigquery>=3.13.0
google-cloud-aiplatform>=1.38.0
numpy>=1.26.0
pyyaml>=6.0.0
lxml>=5.0.0
